Fix bug - world jumping around on every ship move
